Make form less special
