--- src/components/PlayerForm.css
@@ -1,8 +1,0 @@
-.PlayerForm label {
-  display: grid;
-  gap: 4px;
-}
-
-.PlayerForm input {
-  padding: 4px;
-}
