--- src/components/PlayerForm.js
@@ -1,26 +1,0 @@
-import './PlayerForm.css'
-import PropTypes from 'prop-types'
-
-PlayerForm.propTypes = {
-  onSubmit: PropTypes.func.isRequired,
-}
-
-export default function PlayerForm({ onSubmit }) {
-  return (
-    <form className="PlayerForm" onSubmit={handleSubmit}>
-      <label>
-        Add player:
-        <input name="name" type="text" placeholder="Player name" />
-      </label>
-    </form>
-  )
-
-  function handleSubmit(event) {
-    event.preventDefault()
-    const form = event.target
-    const input = form.elements.name
-    onSubmit(input.value)
-    form.reset()
-    input.focus()
-  }
-}
