--- src/components/PlayerForm.stories.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import PlayerForm from './PlayerForm'
-
-export default {
-  title: 'PlayerForm',
-  component: PlayerForm,
-}
-
-export const DefaultForm = args => <PlayerForm {...args} />
